+++ FirmaYonetim/Controllers/ActivityTypeController.cs
@@ -19,31 +19,37 @@ namespace FirmaYonetim.Controllers
         {
             if (Session["user"] == null) return RedirectToAction("Index", "Login");
 
-            ViewModel model = new ViewModel();
-            model.user = PublicFunctions.getUser(conn, Session["user"].ToString());
-            return View(model);
+            return View(new ViewModel()
+            {
+                user = PublicFunctions.getUser(conn, Session["user"].ToString())
+            });
         }
         public ActionResult Detail(Guid? Id)
         {
             if (Session["user"] == null) return RedirectToAction("Index", "Login");
             if (!Id.HasValue) return Redirect("/Activity");
 
+            User user = PublicFunctions.getUser(conn, Session["user"].ToString());
+
             conn.Open();
-            ActivityType activityType = conn.Query<ActivityType>("SELECT * FROM ActivityType WHERE Id = @Id", new ActivityType() { Id = Id }).FirstOrDefault();
+            ActivityType activityType = conn.Query<ActivityType>("SELECT * FROM ActivityType WHERE Id = @Id and CreatedByUserId = @CreatedByUserId", new ActivityType() { Id = Id, CreatedByUserId = user.Id }).FirstOrDefault();
             conn.Close();
 
-            ViewModel model = new ViewModel();
-            model.activityType = activityType;
-            model.user = PublicFunctions.getUser(conn, Session["user"].ToString());
-            return View(model);
+            return View(new ViewModel()
+            {
+                activityType = activityType,
+                user = PublicFunctions.getUser(conn, Session["user"].ToString())
+            });
         }
         public ActionResult Delete(Guid? Id)
         {
             if (Session["user"] == null) return RedirectToAction("Index", "Login");
             if (!Id.HasValue) return Redirect("/Activity");
 
+            User user = PublicFunctions.getUser(conn, Session["user"].ToString());
+
             conn.Open();
-            conn.Execute("DELETE FROM ActivityType WHERE Id = @Id", new ActivityType() { Id = Id });
+            conn.Execute("DELETE FROM ActivityType WHERE Id = @Id and CreatedByUserId = @CreatedByUserId", new ActivityType() { Id = Id, CreatedByUserId = user.Id });
             conn.Close();
 
             return Redirect("/Activity");
@@ -55,8 +61,10 @@ namespace FirmaYonetim.Controllers
         {
             if (Session["user"] == null) return RedirectToAction("Index", "Login");
 
+            User user = PublicFunctions.getUser(conn, Session["user"].ToString());
+
             conn.Open();
-            conn.Execute("INSERT INTO ActivityType (Text) VALUES (@Text)", new ActivityType() { Text = text });
+            conn.Execute("INSERT INTO ActivityType (Text, CreatedByUserId) VALUES (@Text, CreatedByUserId)", new ActivityType() { Text = text, CreatedByUserId = user.Id });
             conn.Close();
 
             return Redirect("/Activity");
@@ -66,8 +74,10 @@ namespace FirmaYonetim.Controllers
         {
             if (Session["user"] == null) return RedirectToAction("Index", "Login");
 
+            User user = PublicFunctions.getUser(conn, Session["user"].ToString());
+            
             conn.Open();
-            conn.Execute("UPDATE ActivityType SET Text = @Text WHERE Id=@Id", new ActivityType() { Id = Id, Text = text });
+            conn.Execute("UPDATE ActivityType SET Text = @Text WHERE Id=@Id and CreatedByUserId=@CreatedByUserId", new ActivityType() { Id = Id, Text = text, CreatedByUserId = user.Id });
             conn.Close();
 
             return Redirect("/ActivityType/Detail/" + Id);
